requirements.txt
fastapi
torch
transformers
uvicorn
accelerate>=0.26.0
python-multipart
python-dotenv